(#9983) Use binary mode when reading and writing FileBucketFiles
Previously, the Puppet::FileBucketFile terminus used text mode in its
save and find methods. It also used text mode when validating that a
file that it was asked to save matched what was already in the file
bucket.

This commit changes these methods to use binary mode, which is a noop
on Unix.