Fix failing CA Interface specs on Ruby 1.9
These tests were creating unnecessary stub objects rather than making
actual Certificate and CertificateRequests. Unsurprisingly, they were
thus failing to stub a method on Ruby 1.9 which was not needed in 1.8.
This commit changes them to use actual objects, to both correct this
problem and improve their robustness for the future.