(#9544) Stub command in package spec that needs root priviledges
The portage package provider on gentoo uses the helper program eix to
query packages.  Eix uses a cache and the portage provider calls
eix-update in case the cache is out of date.  The eix-update needs root
priviledges (or the user at least has to be in the portage group) that
may not be the case when running the specs.

Stub the update_eix method to prevent puppet from modifying the current
system when running the specs.