Configure rcov and flay tests for metric_fu.
Currently if you run rake metrics:all on puppet, you get no results
from flay (because it needs to be told the directory to look in),
and incomplete results from rcov (because it needs to be pointed to rspec).
This patch adds the necessary configuration to the rake task.
The metric_fu gem will need to be installed for the metrics:all task to exist.