(#9109) Retrieve request parameters from the request body for POSTs
When using Puppet with Mongrel, the facts were being lost with 2.7.0+
clients, since they were switched over to using POST requests for fact
submission.  This was happening because the request parameters were
only being retrieved from the query parameters of the URL, which will
not include the POSTed data.

We now merge the body of the post together with the query parameters
when dealing with POST requests.

We also rewind the request body after reading it, since retrieving the
request body multiple times would return an empty string after the
first time.

Paired-with: Nick Lewis <nick@puppetlabs.com>