(#10269) Search bit not set on newly created directories
Previously, puppet would not set the user, group, or other
execute/search bits for newly created directories. This was a
regression introduced in 2.7.6 in commit
22bfd9ce83ff52d909a43c370ba71112ed4961a4.

This was caused because the dirmask'ing was occurring when munging the
mode property, but that was occurring prior to the ensure property
being synced, which is when the directory is actually created. And
since the directory did not exist, the executable bits were never
added.

Similar failures would occur if the path referred to a non-executable
file or link that we wanted to change to a directory.

This commit reverts the changes to the dirmask and munge methods, and
moves the call to dirmask back to the retrieve method. This way we can
be sure that the directory has been created by the time we call
dirmask.

Ideally, we could know at munge time whether we are going to create a
directory as opposed to a file or link. But the logic for that depends
on many other properties, e.g. source, target, etc. The easiest thing
is to just revert the change.

Paired-with: Nick Lewis <nick@puppetlabs.com>