(#10315) Add pending tests when following symlinks
While fixing #10269, I discovered a regression originally filed as
2.6.8 while fixing #6856 for managed symlinks.

When puppet is syncing a file resource, whose source is a symlink that
points to a readable directory, it does not set the execute bit on the
newly created directory.

Conversely, if the symlink points to a directory that is not readable,
puppet does set the execute bit on the newly created directory, when
it shouldn't.

Another problem caused by the change for #6856, is when puppet is
syncing a file resource, whose source is a symlink to a file, puppet
does not create the file.

This commit does not fix these issues, but adds pending tests for
them.