(#10365) Add pending test when file overwrites an executable directory
While fixing #10269, I discovered an issue that has been broken since
at least 0.23.0, which is if puppet manages a file resource that
should not be executable, but an existing executable directory exists,
it will make the newly created file executable. Fortunately, the next
time puppet runs it will remove the executable bits from the file.

This commit does not fix this bug, but adds a pending test for it.