(#8255) Always use string modes when creating resources from FileSetting settings
Since we're setting the 'is', rather than the 'should' on the
resource, the property's munge method is not automatically called for
us and we need to make sure we're always passing the stringified
version of the mode to the resource.

By doing this, we fix the problem where 'puppet --genmanifest' would
output the (base 10) integer version of the mode in the generated
manifest for settings where the mode was not specified as a string.

It would have been nice to re-use the munge from the mode property
directly, but there doesn't appear to be a good/clean way to do this
(especially without reaching into the internals of other objects).

Another alternative would have been to modify []= to call munge for
us, but this isn't really a change to be made in a point release,
especially not without very careful thought about the implications of
such a change.