(#9617) Use an RbTreeMap to store ready resources
The operations we need to perform on the ready list (insertion in sorted
order, random deletion, and deletion of the first element) are best
implemented with a sorted set. Because the core Ruby SortedSet class
depends on the C rbtree gem, we use an RbTreeMap instead.

Previously, we would sort the entire ready list each time we needed to
retrieve the next resource. With a large number of non-interdependent
resources, this is quite costly. Now, we keep the list in sorted order,
and simply pull off the next element, resulting in significant
performance gains.

Because we are now indexing resources by their
unguessable_deterministic_key, rather than simply sorting them, we need
to ensure these keys are unique. So we use the resource ref, rather than
the title, when generating the key.

Paired-With: Josh Cooper <josh@puppetlabs.com>