(#10614) Add method for detecting Windows volumes that support ACLs
Added a method to detect whether the root of a given path is on a
volume that supports persistent Windows ACLs. Note it is not enough to
check that the volume type is 'NTFS' as some filesystems claim to be
NTFS, but do not support ACLs, e.g. NetApp. Other filesystems like FAT
and CIFS do not support ACLs either.

This commit does not add any new windows gem dependencies, as the
Windows::Volume module is part of the windows-pr gem.

Paired-with: Nick Lewis <nick@puppetlabs.com>