Fixing #6907 - unsuitable providers no longer fail
We were previously failing in two circumstances:

* Prefetching unsuitable providers at the beginning
  of a transaction
* Choosing a default provider when all available
  providers are unsuitable.

The major downside of these failures is that pluginsync
could not reliably be used to download new providers
and use them in the same run.

This commit fixes those two areas, and tests that they
no longer fail.  It unfortunately does not (and, really,
cannot as a unit test) test that pluginsync works with
this kind of provider, but it gets as close as possible.

Signed-off-by: Luke Kanies <luke@puppetlabs.com>