(#4865) Debug logging when we start evaluating resources.
The `evaltrace` option allowed individual resource evaluation time to be
tracked, which made it easier to post-hoc identify which resources took long
periods of time to process.

It is also helpful, when doing live debugging, to know where the hang happens;
to support that we now log a debug message about starting the evaluation of
the resource before we go into the process.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>