(#9671) Stop tracking size and height of nodes in RbTreeMap
Profiling revealed RbTreeMap#update_size to be one of the most expensive methods
in a Puppet run. The only function of this method is to track the size
and height of every node, which in turn is used to provide the size and
height of the tree.

Individual nodes aren't accessible, so there is no need to keep track of
their size and height. While height of the tree is theoretically
interesting, we currently don't use it anywhere, and are unlikely to in
the future.

So we simplified the size tracking to simply update the size of the tree
on insertion and deletion.

Paired-With: Josh Cooper <josh@puppetlabs.com>