(#9671) Implement RbTreeMap#each recursively, and #first/#last explicitly
This code was using a complex iterative approach, when a simpler
recursive one would suffice. Because the tree is balanced and the height
is tightly bounded, there is no risk of a stack overflow until the size
of the tree grows to around 10^1310 nodes.

Additionally, #first and #last are extremely important operations when
using the tree as a queue. The current implementation of #first, granted
by Enumerable, is quite slow. We already have a method (#min_recursive)
which retrieves the first node, so implementing #first is trivial.
Similarly, #last is not provided by Enumerable, so we implement that as
well.

Paired-With: Josh Cooper <josh@puppetlabs.com>