(#9671) Exit early from #eval_generate if nothing is created
We now exit early if #eval_generate does not produce any resources.
This has the impact of not adding a completed_ whit for the resource,
which is not necessary without any children.

We also now return true if #eval_generate created resources, and false
if it did not. This allows callers to reason about whether the graph
actually changed.