(#9671) Generated resources should not depend on the completed_ whit
In #eval_generate, we rewire anything which depends on the resource to instead
depend on its completed_ whit, which signifies the completion of the "whole"
resource (the resource + its children). This ensures anyting depending on the
resource doesn't run until the resource and its children have all run. In some
cases (such as when generated resources autorequired the generating resource)
we would also rewire the child resources to depend on the completed_ whit.

This is incorrect, because we consider child resources to be "part" of the
resource; that is, the completed_ whit should not be 'done' until the children
are done. Now, when rewiring these dependencies, we skip generated resources.