(#7500) Stop pw provider using -p
The previous commit enabled support for setting passwords, but it
didn't deal with what happens when an account is created. The existing
code takes the first character of each property and uses it as a flag.
However, with pw, the -p flag is for setting the password age.

The result is that the password isn't set at create time and that the
password is set to expire. The next run of puppet correctly sets the
password but the expiry is still set.

This fix comes from bug #7500.

Test written by me.