(#6412) Return :undef when accessing non-existing hash/array elements
The truethiness test done in P::P::Scope.true? doesn't handle nil
values, because it is not valid as result of a puppet expression.
Unfortunately ruby hash['invalid'] or array[inexistant_index] returns
nil which then was considered as true in an if expression.
This patch makes sure :undef is returned in those cases.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>