(#10676) Include all resource statuses in reports regardless of count
Before, we would only include resource status states, and
failure/success counts if there were resources that were in these
various states.  We now initialize the hashes used to gather the
counts of these metrics so that they will include all known states,
even if the count of resources that were in that state is zero.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>
Signed-off-by: Jacob Helwig <jacob@puppetlabs.com>