(#7106) Obey specified owner, group, and permissions for last run summary file
When writing the last run summary (puppet settings lastrunfile), we
now enforce any specified mode.  The default permissions have also
been changed to 0644.  This can be changed in puppet.conf:

  [agent]
  ...
  lastrunfile = /tmp/lastrun.yml { mode=0664 }
  ...

While the owner and group can be specified, the only choices that work
at this time are root, and the service user/group.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>
Signed-off-by: Jacob Helwig <jacob@puppetlabs.com>