maint: Fix failing specs for Windows exec provider
These specs were trying to stub the microsoft_windows feature to enable
the Windows exec provider for the test. However, the Windows provider
was changed in ad98d47 to use the operatingsystem fact for its confine,
rather than the feature, so these failed. This changes them to also stub
the operatingsystem fact.

Reviewed-by: Josh Cooper <josh@puppetlabs.com>