(#10887) Wait until puppet is stopped before returning
In some cases, the agent init script will return off a stop before the puppet
agent has actually stopped. This patch verifies that the process has stopped by
looping on `kill -0` on the pid number (if it exists) until kill -0 returns 1,
at which point the process will really have stopped. After 60
seconds/iterations, it falls back to a kill -9 on the pid number. This patch
addresses that concern with the redhat agent init script.

Signed-off-by: Matthaus Litteken <matthaus@puppetlabs.com>