(#10111) - OCSP server checks
This commit adds client certificate OCSP verification in the server.
This makes sense only if the master is not a CA, in which case checking
the CRL directly would be way faster.

To do this the server needs to have access to the client certificate.
This is not a problem for webrick based setups because puppet is the
SSL end point.

On Mongrel and Passenger setups, the SSL endpoint needs to forward
the client certificate to Puppet.

On nginx in proxy mode this can be done with:
proxy_set_header    X-Client-Certificate     $ssl_client_cert;

On apache/passenger setups this can be done with:
SetEnv HTTP_X_CLIENT_CERTIFICATE %{SSL_CLIENT_CERTIFICATE}e
or
SSLOptions +StdEnv +ExportCertData
(but the puppet option ssl_client_certificate_header should be changed)

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>