Restore mixin behaviours to indirected classes.
The indirector once again functions directly as a mixin to classes it extends,
allowing direct and explicit invocation of indirection behaviour through the
model class.

It also restores the instance behaviour of allowing direct save working
through an instance of the model, without having to contort through the
typical hoop of:

    facts.class.indirection.save(foo)
    # ...or: Puppet::Node::Facts.indirection.save(foo)

    # compare:
    facts.save

It also removes the extra hoop of `.indirection` from every invocation of the
basic indirected operations, `find`, `save`, `search`, `destroy`, and `head`.

This is done, ultimately, because there was extra cognitive and code load
caused by the cleanup that was never paid back by seeing the remained of the
changes pushed into the code.

It also causes a constant friction between the 2.6 series, where this was not
done, and the 2.7 release series, where it was.

If the cleanup is resumed then this single, atomic change can easily be
reverted; in what seems the more likely case that any cleanup takes another
form after so long, this seems to have no lasting pain.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>