Forward `terminus` to the indirection from the model.
One of the common operations interacting with the indirection subsystem is to
request a concrete terminus for a specific model class.  Instead of requiring
this to manually go through the indirection instance we forward the method
internally.

This removes a tiny section of ceremony from the system, and gives an equally
clean and transparent interface to the model class.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>