Forward terminus and cache class manipulation from the model.
For better or worse, one of the most common operations on the indirection
class is to manipulate the current terminus and cache class settings.

This change unifies that with direct operations, bringing those changes into
the model.  This reduces the ceremony around that manipulation, and makes
clear that the changes are local.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>