`reset_terminus_class` is redundant to just assigning `nil`
Resetting the cache class for an indirection was done by simply assigning
`nil` to the field; doing the same for the actual terminus required going
through a separate method, `reset_terminus_class`.

Aside from contorting some code to carefully detect a `nil` value and avoid
resetting the terminus in a way that raised errors, this added literally no
value to the system - there was no semantic difference, and there were
multiple different ways of achieving the same result on different parts of the
same object.

So, this collapses the extra method into a simple `nil` assignment; that leads
to simpler code and less ceremony for dealing with the indirections.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>