(#7656) Port PMT to faces, bundle in Puppet core
Before this patch, Puppet did not have the ability to interact with the
Puppet Module Forge (PMF).

This patch is one of many patches that add a new module_tool face to
core Puppet. This patch provides the foundation bits which setup the
ground work for adding additional actions and features.