(#7656) Add changes action to module_tool face
Before this patch, the module_tool face did not have the ability to
show modified files of an installed module, and there was no test
coverage for the module_tool changes action.

This patch adds support for showing modified files of an installed
module to the module_tool face.

Example usage:

    $ puppet module_tool changes /path/to/module

This patch includes unit tests for the module_tool changes action.